+++ jirust-cli/src/runners/jira_cmd_runners/project_cmd_runner.rs
@@ -4,7 +4,9 @@ use jira_v3_openapi::apis::configuration::Configuration;
 use jira_v3_openapi::apis::issues_api::{
     get_create_issue_meta_issue_type_id, get_create_issue_meta_issue_types,
 };
-use jira_v3_openapi::apis::projects_api::search_projects;
+use jira_v3_openapi::apis::projects_api::{create_project, search_projects};
+use jira_v3_openapi::models::create_project_details::{AssigneeType, ProjectTypeKey};
+use jira_v3_openapi::models::{CreateProjectDetails, ProjectIdentifiers};
 use jira_v3_openapi::models::{
     FieldCreateMetadata, IssueTypeIssueCreateMetadata, project::Project,
 };
@@ -53,6 +55,70 @@ impl ProjectCmdRunner {
         config.base_path = cfg_file.get_jira_url().to_string();
         config.basic_auth = Some((auth_data.0, Some(auth_data.1)));
         ProjectCmdRunner { cfg: config }
+    }
+
+    /// Create a new Jira project using the provided parameters.
+    ///
+    /// # Arguments
+    /// * `params` - The parameters for creating the project.
+    ///
+    /// # Returns
+    /// A `Result` containing the project identifiers if successful, or an error if failed.
+    ///
+    /// # Examples
+    ///
+    /// ```no_run
+    /// use jirust_cli::runners::jira_cmd_runners::project_cmd_runner::{ProjectCmdRunner, ProjectCmdParams};
+    /// use jirust_cli::config::config_file::ConfigFile;
+    /// use toml::Table;
+    ///
+    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
+    /// # tokio_test::block_on(async {
+    /// let cfg_file = ConfigFile::new("auth_token".to_string(), "jira_url".to_string(), "standard_resolution".to_string(), "standard_resolution_comment".to_string(), Table::new());
+    /// let project_cmd_runner = ProjectCmdRunner::new(cfg_file);
+    ///
+    /// let mut params = ProjectCmdParams::new();
+    /// params.project_key = Some("TEST".to_string());
+    /// params.project_name = Some("Test Project".to_string());
+    /// params.project_description = Some("This is a test project".to_string());
+    /// params.project_field_configuration_id = Some(12345);
+    /// params.project_issue_security_scheme_id = Some(67890);
+    /// params.project_issue_type_scheme_id = Some(54321);
+    ///
+    /// let projects = project_cmd_runner.create_jira_project(params).await?;
+    ///
+    /// # Ok(())
+    /// # })
+    /// # }
+    /// ```
+    pub async fn create_jira_project(
+        &self,
+        params: ProjectCmdParams,
+    ) -> Result<ProjectIdentifiers, Box<dyn std::error::Error>> {
+        let mut project_data = CreateProjectDetails::new(
+            params.project_key.expect("Project key is required!"),
+            params.project_name.expect("Project name is required!"),
+        );
+        project_data.description = params.project_description;
+        project_data.field_configuration_scheme = params.project_field_configuration_id;
+        project_data.issue_security_scheme = params.project_issue_security_scheme_id;
+        project_data.issue_type_scheme = params.project_issue_type_scheme_id;
+        project_data.issue_type_screen_scheme = params.project_issue_type_screen_scheme_id;
+        project_data.notification_scheme = params.project_notification_scheme_id;
+        project_data.permission_scheme = params.project_permission_scheme_id;
+        project_data.workflow_scheme = params.project_workflow_scheme_id;
+        project_data.lead_account_id = params.project_lead_account_id;
+        project_data.assignee_type = if params
+            .project_assignee_type
+            .unwrap_or("unassigned".to_string())
+            == "lead"
+        {
+            Some(AssigneeType::ProjectLead)
+        } else {
+            Some(AssigneeType::Unassigned)
+        };
+        project_data.project_type_key = Some(ProjectTypeKey::Software);
+        Ok(create_project(&self.cfg, project_data).await?)
     }
 
     /// Lists Jira projects.
@@ -221,6 +287,17 @@ impl ProjectCmdRunner {
 pub struct ProjectCmdParams {
     pub project_key: Option<String>,
     pub project_issue_type: Option<String>,
+    pub project_name: Option<String>,
+    pub project_description: Option<String>,
+    pub project_field_configuration_id: Option<i64>,
+    pub project_issue_security_scheme_id: Option<i64>,
+    pub project_issue_type_scheme_id: Option<i64>,
+    pub project_issue_type_screen_scheme_id: Option<i64>,
+    pub project_notification_scheme_id: Option<i64>,
+    pub project_permission_scheme_id: Option<i64>,
+    pub project_workflow_scheme_id: Option<i64>,
+    pub project_lead_account_id: Option<String>,
+    pub project_assignee_type: Option<String>,
     pub projects_page_size: Option<i32>,
     pub projects_page_offset: Option<i32>,
 }
@@ -249,6 +326,17 @@ impl ProjectCmdParams {
         ProjectCmdParams {
             project_key: None,
             project_issue_type: None,
+            project_name: None,
+            project_description: None,
+            project_field_configuration_id: None,
+            project_issue_security_scheme_id: None,
+            project_issue_type_scheme_id: None,
+            project_issue_type_screen_scheme_id: None,
+            project_notification_scheme_id: None,
+            project_permission_scheme_id: None,
+            project_workflow_scheme_id: None,
+            project_lead_account_id: None,
+            project_assignee_type: None,
             projects_page_size: None,
             projects_page_offset: None,
         }
@@ -275,9 +363,20 @@ impl From<&ProjectArgs> for ProjectCmdParams {
     /// use jirust_cli::args::commands::{ProjectArgs, ProjectActionValues, PaginationArgs, OutputArgs};
     ///
     /// let project_args = ProjectArgs {
-    ///     project_act: ProjectActionValues::List,
+    ///     project_act: ProjectActionValues::GetIssueTypeFields,
     ///     project_key: Some("project_key".to_string()),
     ///     project_issue_type: Some("project_issue_type".to_string()),
+    ///     project_name: None,
+    ///     project_description: None,
+    ///     project_field_configuration_id: None,
+    ///     project_issue_security_scheme_id: None,
+    ///     project_issue_type_scheme_id: None,
+    ///     project_issue_type_screen_scheme_id: None,
+    ///     project_notification_scheme_id: None,
+    ///     project_permission_scheme_id: None,
+    ///     project_workflow_scheme_id: None,
+    ///     project_lead_account_id: None,
+    ///     project_assignee_type: None,
     ///     pagination: PaginationArgs { page_size: Some(10), page_offset: None },
     ///     output: OutputArgs { output_format: None, output_type: None },
     /// };
@@ -293,6 +392,17 @@ impl From<&ProjectArgs> for ProjectCmdParams {
         ProjectCmdParams {
             project_key: value.project_key.clone(),
             project_issue_type: value.project_issue_type.clone(),
+            project_name: value.project_name.clone(),
+            project_description: value.project_description.clone(),
+            project_field_configuration_id: value.project_field_configuration_id,
+            project_issue_security_scheme_id: value.project_issue_security_scheme_id,
+            project_issue_type_scheme_id: value.project_issue_type_scheme_id,
+            project_issue_type_screen_scheme_id: value.project_issue_type_screen_scheme_id,
+            project_notification_scheme_id: value.project_notification_scheme_id,
+            project_permission_scheme_id: value.project_permission_scheme_id,
+            project_workflow_scheme_id: value.project_workflow_scheme_id,
+            project_lead_account_id: value.project_lead_account_id.clone(),
+            project_assignee_type: value.project_assignee_type.clone(),
             projects_page_size: value.pagination.page_size,
             projects_page_offset: Some(
                 i32::try_from(value.pagination.page_offset.unwrap_or(0))